+++ include/loans.h
@@ -1,0 +1,12 @@
+#ifndef LOANS_H
+#define LOANS_H
+
+typedef struct loans
+{
+    char name[50];
+    char cpf[11];
+    char title[50];
+    int id;
+} Loans;
+
+#endif
