+++ src/main.c
@@ -3,9 +3,11 @@
 #include <stdlib.h> //memory manipulation or system
 #include <stddef.h> //definition of null and size
 #include <locale.h> //set language
-#include "../include/structs.h"
 #include "../include/system.h"
-#include "../include/functions.h"
+#include "../include/home.h"
+#include "../include/user.h"
+#include "../include/book.h"
+#include "../include/loans.h"
 
 int main()
 {
